chore[bench-website]: add back tpc-ds to query_bench
