+++ bench-vortex/src/tpch/mod.rs
@@ -18,7 +18,6 @@ use vortex::{ArrayRef, IntoArray};
 use crate::{BenchmarkDataset, datasets};
 
 pub mod dbgen;
-pub mod duckdb;
 pub mod schema;
 pub mod tpch_benchmark;
 pub mod tpchgen;
@@ -104,7 +103,7 @@ pub async fn register_vortex_compact_file(
 }
 
 /// Load a table as an uncompressed Vortex array.
-pub async fn load_table(data_dir: impl AsRef<Path>, name: &str, schema: &Schema) -> ArrayRef {
+pub async fn load_table(data_dir: impl AsRef<Path>, name: &str, schema: Schema) -> ArrayRef {
     // Create a local session to load the CSV file from the path.
     let path = data_dir.as_ref().join(name).with_extension("csv");
     let record_batches = SessionContext::new()
@@ -114,7 +113,7 @@ pub async fn load_table(data_dir: impl AsRef<Path>, name: &str, schema: &Schema)
                 .delimiter(b'|')
                 .has_header(false)
                 .file_extension("csv")
-                .schema(schema),
+                .schema(&schema),
         )
         .await
         .unwrap()
